refactor: ui folder removed at modules
