+++ frontend/src/entities/player/PlayerStatsCard.styles.ts
