--- frontend/src/entities/player/ui/index.ts
@@ -1,1 +1,0 @@
-export { PlayerStatsCard } from './PlayerStatsCard'
