+++ frontend/src/entities/team/TeamCard/TeamCard.props.ts
