+++ frontend/src/entities/team/TeamCard/TeamCard.styles.ts
