+++ frontend/src/entities/team/TeamCard/index.ts
