+++ frontend/src/entities/team/TeamStatsCard/TeamStatsCard.styles.ts
