+++ frontend/src/entities/team/TeamStatsCard/index.ts
