--- frontend/src/entities/team/ui/index.ts
@@ -1,2 +1,0 @@
-export { TeamStatsCard } from './TeamStatsCard'
-export { TeamCard } from './TeamCard'
