+++ frontend/src/pages/MainPage/index.ts
@@ -1,1 +1,1 @@
-export { MainPage } from './ui'
+export { MainPage } from './MainPage'
