--- frontend/src/pages/MainPage/ui/index.ts
@@ -1,1 +1,0 @@
-export { MainPage } from './MainPage'
