+++ frontend/src/ui/StatsTable/StatsTable.props.ts
