+++ frontend/src/ui/StatsTable/StatsTable.styles.ts
