+++ frontend/src/ui/StatsTable/StatsTable.tsx
