--- frontend/src/ui/StatsTable/ui/index.ts
@@ -1,1 +1,0 @@
-export { StatsTable } from './StatsTable'
