+++ frontend/src/widgets/MainHeader/MainHeader.styles.ts
