--- frontend/src/widgets/MainHeader/ui/index.ts
@@ -1,1 +1,0 @@
-export { MainHeader } from './MainHeader'
