+++ frontend/src/widgets/TeamsList/TeamsList.styles.ts
