+++ frontend/src/widgets/TeamsList/index.ts
@@ -1,1 +1,1 @@
-export { TeamsList } from './ui'
+export { TeamsList } from './TeamsList'
