--- frontend/src/widgets/TeamsList/ui/index.ts
@@ -1,1 +1,0 @@
-export { TeamsList } from './TeamsList'
